requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
spacy
transformers
torch
dateparser
pydantic
